Arrumando erros do SonarQube (#2) [Vitor, Arthur]

+++ src/main/java/br/usp/ime/cogroo/util/CriptoUtils.java
@@ -70,7 +70,7 @@ public final class CriptoUtils {
 		StringBuffer buf = new StringBuffer();
 
 		for (int i = 0; i < b.length; i++) {
-			int j = ((int) b[i]) & 0xFF;
+			int j = (b[i]) & 0xFF;
 			buf.append(hexDigits.charAt(j / 16));
 			buf.append(hexDigits.charAt(j % 16));
 		}
